+++ frontend/src/components/dashboard/ProfileCard.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Button } from "@nextui-org/react";
 import ProfilePictureAvatar from "@/components/common/ProfilePictureAvatar";
-import { UserService } from "@/helpers/user/api_wrappers";
+import { UserService } from "@/helpers/user/user_api_wrappers";
 import { useRouter } from "next/navigation";
 import { CLIENT_ROUTES } from "@/common/constants";
